Merge master into wpe enhanced

Approved-by: Brett Meehan <bmeehan@atlassian.com>

+++ ami/scripts/init.d/atl-init-synchrony.sh
@@ -259,11 +259,13 @@ function stopSynchrony() {
 function installConfluence {
     atl_log "Checking if ${ATL_CONFLUENCE_SHORT_DISPLAY_NAME} has already been installed"
 
-    atl_log "Creating file /etc/ld.so.conf.d/confluence.conf"
-    echo /usr/lib/jvm/jre-1.7.0-openjdk.x86_64/lib/amd64/server/ > /etc/ld.so.conf.d/confluence.conf
-    sudo ldconfig
-    service collectd restart
-    atl_log "Creating file /etc/ld.so.conf.d/confluence.conf ==> done"
+    if [ "true" = "${ATL_USE_COLLECTD}" ] && [ -e /etc/init.d/collectd ]; then
+        atl_log "Creating file /etc/ld.so.conf.d/confluence.conf"
+        echo /usr/lib/jvm/jre-1.7.0-openjdk.x86_64/lib/amd64/server/ > /etc/ld.so.conf.d/confluence.conf
+        sudo ldconfig
+        if [ -n $ATL_STARTCOLLECTD == "true" ]; then service collectd restart
+        atl_log "Creating file /etc/ld.so.conf.d/confluence.conf ==> done"
+    fi
 
     if [[ -d "${ATL_CONFLUENCE_INSTALL_DIR}" ]]; then
         local ERROR_MESSAGE="${ATL_CONFLUENCE_SHORT_DISPLAY_NAME} install directory ${ATL_CONFLUENCE_INSTALL_DIR} already exists - aborting installation"
@@ -272,8 +274,7 @@ function installConfluence {
     fi
 
     atl_log "Downloading ${ATL_CONFLUENCE_SHORT_DISPLAY_NAME} ${ATL_CONFLUENCE_VERSION} from ${ATL_CONFLUENCE_INSTALLER_DOWNLOAD_URL}"
-    if ! curl -L -f --silent "${ATL_CONFLUENCE_INSTALLER_DOWNLOAD_URL}" -o "$(atl_tempDir)/installer" >> "${ATL_LOG}" 2>&1
-    then
+    if ! curl -L -f --silent "${ATL_CONFLUENCE_INSTALLER_DOWNLOAD_URL}" -o "$(atl_tempDir)/installer" >> "${ATL_LOG}" 2>&1 ; then
         local ERROR_MESSAGE="Could not download installer from ${ATL_CONFLUENCE_INSTALLER_DOWNLOAD_URL} - aborting installation"
         atl_log "${ERROR_MESSAGE}"
         atl_fatal_error "${ERROR_MESSAGE}"
